fix: remove input img link, 'cause it may have cors problem
