+++ src/routes/Home/index.tsx
@@ -1,5 +1,5 @@
 import { generateImageEmbeddings, getDetails, getOssCredentials } from '@/request';
-import { ChangeEvent, useRef, useState } from 'react';
+import { ChangeEvent, SetStateAction, useRef, useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import DisplayImg from '@/components/DisplayImage';
 import { example } from '@/components/constant';
@@ -13,20 +13,19 @@ const Home = () => {
   const [embedding, setEmbedding] = useState<string>(example.embedding);
   const [loading, setLoading] = useState<boolean>(false);
   const [url, setUrl] = useState<string>(example.url);
-  const type = useRef<ImageSourceType>('link');
 
   const learnImg = () => {
     if (loading) return;
     if (url === example.url) {
       toast({
-        content: '请先输入图片地址或者上传本地图片！',
+        content: '请先上传本地图片！',
         type: 'error'
       })
       return;
     }
     const fileName = uuidv4();
     setLoading(true);
-    generateImageEmbeddings({ url: type.current === 'local' ? url.split(',')[1] : url, fileName, type: type.current }).then(({ data }: any) => {
+    generateImageEmbeddings({ url: url.split(',')[1], fileName }).then(({ data }: any) => {
       setEmbedding(data.url);
     }).finally(() => {
       setLoading(false);
@@ -36,7 +35,6 @@ const Home = () => {
   const handleUpload = async (e: ChangeEvent<HTMLInputElement>) => {
     const file = e.target?.files?.[0];
     if (file) {
-      type.current = 'local';
       convertImageFileToBase64(file).then((base64) => {
         setUrl(base64);
       })
@@ -58,21 +56,19 @@ const Home = () => {
         }
       </div>
       <div className='mt-8 text-center'>
-        <div className='flex items-center mb-6'>
-          <input type="text" className='input input-bordered input-primary w-full max-w-xs' placeholder='Input the image link' />
-          <button
-            className='btn btn-primary relative ml-4'
-          >
-            <CloudUploadOutlined className='mr-2 text-lg' />
-            上传图片
-            <input
-              multiple={false}
-              type="file"
-              accept='image/*'
-              onChange={handleUpload}
-              className="w-full opacity-0 absolute block h-full z-0 cursor-pointer"
-            />
-          </button></div>
+        <button
+          className='btn btn-primary relative mr-4'
+        >
+          <CloudUploadOutlined className='mr-2 text-lg' />
+          上传图片
+          <input
+            multiple={false}
+            type="file"
+            accept='image/*'
+            onChange={handleUpload}
+            className="w-full opacity-0 absolute block h-full z-0 cursor-pointer"
+          />
+        </button>
         <button
           type="button"
           className={`text-sm font-semibold text-gray-800`}
