+++ src/utils/clientUtil.ts
@@ -26,3 +26,17 @@ export function convertImageFileToBase64(file: File): Promise<string> {
     };
   });
 }
+
+export async function imageUrlToBase64Client(url: string): Promise<string> {
+  const response = await fetch(url);
+  const blob = await response.blob();
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onloadend = () => {
+      const base64Data = reader.result?.toString().split(',')[1] || '';
+      resolve(base64Data as string);
+    };
+    reader.onerror = reject;
+    reader.readAsDataURL(blob);
+  });
+}
